skill-mind-ai/backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-socketio
python-dotenv
werkzeug
PyPDF2
google-generativeai
